requirements.txt
aiogram==3.10.0
APScheduler==3.10.4
Flask==3.0.3
SQLAlchemy==2.0.23
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
feedparser==6.0.10
pytest